PostViewModelをインタフェースと実装に分ける
