+++ app/src/main/java/com/example/foos/ui/view/screen/post/PostScreen.kt
@@ -40,7 +40,7 @@ import kotlinx.coroutines.launch
 @OptIn(ExperimentalPermissionsApi::class)
 @Composable
 fun PostScreen(
-    viewModel: PostViewModel,
+    viewModel: PostViewModelImpl,
     navController: NavController,
     sharedViewModel: ScreenViewModel,
 ) {
